#100.16 Test registración y login con email #102.26 y #108.20 Tipo de publicacion en Publicar y crear Alerta #104.9 La distancia la vuelvo a pasar a double, como en el server #116.1 PublicacionAtributos Singleton

+++ app/src/main/java/com/tdp2grupo9/fragment/adopcion/ResultadosBusquedaFragment.java
@@ -71,7 +71,7 @@ public class ResultadosBusquedaFragment extends PublicacionesFragment {
         publicacion.setCompatibleCon(getCompatibleConPublicacion(arguments.getInt("compatiblecon", 0)));
         publicacion.setVacunasAlDia(getVacuntasAlDiaPublicacion(arguments.getInt("vacunas", 0)));
         publicacion.setPapelesAlDia(getPapelesAlDiaPublicacion(arguments.getInt("papeles", 0)));
-        publicacion.setDistancia(getDistanciaPublicacion(arguments.getInt("distancia", -1)));
+        publicacion.setDistancia(getDistanciaPublicacion(arguments.getDouble("distancia", -1.0)));
         publicacion.setLongitud(getLongitudPublicacion(arguments.getDouble("longitud", -1.0)));
         publicacion.setLatitud(getLatitudPublicacion(arguments.getDouble("latitud", -1.0)));
 
@@ -91,8 +91,8 @@ public class ResultadosBusquedaFragment extends PublicacionesFragment {
         return latitud;
     }
 
-    private Integer getDistanciaPublicacion(int distancia) {
-        if (distancia == -1)
+    private Double getDistanciaPublicacion(double distancia) {
+        if (distancia == -1.0)
             return null;
         return distancia;
     }
